Bug 73298 - Sync Interval - Default vs. Individual folder
There was a problem with the types used in the math so bad values were showing up


git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk/ifolder@4978 db2431b9-df06-0410-971c-a38ea9a75485

+++ src/OSXClient/src/iFolderPrefsController.m
@@ -108,7 +108,7 @@ static iFolderPrefsController *prefsSharedInstance = nil;
 		[syncValue setEnabled:YES];
 		[syncLabel setEnabled:YES];
 
-		int value = [TimeSpan getTimeSpanValue:syncInterval];
+		long value = [TimeSpan getTimeSpanValue:syncInterval];
 		[syncValue setStringValue:[NSString stringWithFormat:@"%d", value]];
 		[syncUnits selectItemAtIndex:[TimeSpan getTimeIndex:syncInterval] ];
 	}
@@ -117,7 +117,7 @@ static iFolderPrefsController *prefsSharedInstance = nil;
 
 -(void)saveCurrentSyncInterval
 {
-	int seconds;
+	long seconds;
 	if([enableSync state])
 		seconds = [TimeSpan getSeconds:[syncValue intValue]
 						withIndex:[syncUnits indexOfSelectedItem]];
